+++ CHANGELOG.md
@@ -3,10 +3,15 @@
 
 All notable changes to `file-url-signer` will be documented in this file
 
+# 1.0.2
+
+- File Entity class scan and extract parameters from the file name;
+- Fixes to `separator` in File Entity class.
+
 # 1.0.1
 
 - Fixed bugs in `UrlSigner` class;
-- Created `ParameterCollection` and `File` entity classes.
+- Created `ParameterCollection` and `File` entity classes;
 - Simplified `UrlSigner` class;
 - Updated `README.md`.